[FE] REFACT: LoginButtonGroup container와 presenter로 분리 #1739
